Split ScalarTest.cc into multiple files for faster compilation
